requirements.txt
requests==2.31.0
PyYAML==6.0.1
sentence-transformers==3.0.1
chromadb==0.5.3
anthropic==0.40.0
python-dotenv==1.0.1
streamlit==1.41.1
google-api-python-client>=2.140.0
google-auth==2.27.0
